Refactor file org, components
