+++ App.js
@@ -1,41 +1,12 @@
 import { useState } from 'react';
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View, TextInput, Button } from 'react-native';
+import { TextInput } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { Formula, FormulaHelpers } from './vdotCalc';
-
-function HomeScreen({ navigation }) {
-
-  return (
-    <View style={styles.container}>
-      <Text>Home Screen</Text>
-      <Button title="Roster" onPress={() => navigation.navigate('Roster')} />
-      <Button title="Workout" onPress={() => navigation.navigate('Workout')} />
-      <StatusBar style="auto" />
-    </View>
-  );
-}
-
-function RosterScreen() {
-
-  return (
-    <View style={styles.container}>
-      <Text>Roster Screen</Text>
-      <StatusBar style="auto" />
-    </View>
-  );
-}
-
-function WorkoutScreen() {
-  
-  return (
-    <View style={styles.container}>
-      <Text>Workout Screen</Text>
-      <StatusBar style="auto" />
-    </View>
-  );
-}
+import { styles } from './src/constants/styles';
+import { Formula, FormulaHelpers } from './src/utils/vdotCalc';
+import HomeScreen from './src/Components/HomeScreen';
+import RosterScreen from './src/Components/RosterScreen';
+import WorkoutScreen from './src/Components/WorkoutScreen';
 
 const Stack = createNativeStackNavigator();
 
@@ -51,11 +22,3 @@ export default function App() {
     </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  }
-});
