+++ src/utils/vdotCalc.js
